Change master/worker and server/client structure
