+++ .gitignore
@@ -43,5 +43,6 @@ website/static/api
 .metals/
 .bloop/
 */bloop/
+project/metals.sbt
 project/project/metals.sbt
 project/project/project/metals.sbt
