+++ .scalafmt.conf
@@ -1,0 +1,2 @@
+version = "3.5.3"
+runner.dialect = scala213
